+++ railties/lib/rails/application/finisher.rb
@@ -21,10 +21,13 @@ module Rails
 
       initializer :add_builtin_route do |app|
         if Rails.env.development?
-          app.routes.append do
+          app.routes.prepend do
             get '/rails/info/properties' => "rails/info#properties", internal: true
             get '/rails/info/routes'     => "rails/info#routes", internal: true
             get '/rails/info'            => "rails/info#index", internal: true
+          end
+
+          app.routes.append do
             get '/'                      => "rails/welcome#index", internal: true
           end
         end